[Refactor] png icons to svg icons(#61)
